+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
- - Display collected version info function.
+ - Display collected windows version info function.
+ - Display windows version info function.
 
 ### Changed
  - The output user name was changed from UserName to UserDomainName+UserName.