+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Display collected windows version info function.
  - Display windows version info function.
+ - Japanese language support.
 
 ### Changed
  - The output user name was changed from UserName to UserDomainName+UserName.